pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openeye"
version = "1.0.0"
description = "Comprehensive video processing system for drone camera systems"
readme = "README.md"
authors = [
    {name = "OpenEye Team", email = "example@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Rust",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.20.0",
    "opencv-python>=4.8.0",
    "torch>=2.0.1",
    "tensorflow>=2.12.0",
    "fastapi>=0.103.1",
    "uvicorn>=0.23.2",
    "websockets>=11.0.3",
    "pydantic>=2.4.2",
    "langchain>=0.0.267",
    "openai>=1.3.0",
    "aiortc>=1.5.0",
    "av>=9.0.0",
    "pyyaml>=6.0",
    "structlog>=23.1.0",
    "pillow>=10.0.0",
    "cvxpy>=1.3.2",
    "ffmpeg-python>=0.2.0",
    "faiss-cpu>=1.7.4",
    "chromadb>=0.4.13",
    "scipy>=1.10.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-benchmark>=4.0.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "ruff>=0.0.253",
    "sphinx>=7.1.2",
    "sphinx-rtd-theme>=1.3.0",
]
cuda = [
    "cupy-cuda12x>=12.2.0",
    "nvidia-cuda-runtime-cu12>=12.2.0",
    "nvidia-cudnn-cu12>=8.9.3.28",
]
vr = [
    "pyvista>=0.40.0",
    "pyopenvr>=2.0",
]

[tool.setuptools.packages.find]
where = ["src"]